♻️🚚 Refactor `ChatGPTService`.
